CI-62 Failed to correctly setup job.
I think this was caused by a race condition whereby both the job and the importNow() method were trying to create the data and one of them was failing. This is difficult to reproduce but triggering the job at the end should fix it.

Each repository call should be in it’s own transaction which should mean the data is setup and committed to the DB by the time we trigger the import.