CI-70 Make sure to set a SameSite attribute.
With Chrome 80 it will reject cookies if we don’t specify a SameSite attribute and if we don’t do anything then older browsers will also stop working (eg iOS 12) so we set 2 cookies, one for newer browsers and one for older.

The alternative is user agent blacklisting for the SameSite attribute.