CI-75 We weren’t correctly setting the status
When dispatching to the error servlet we need to use request attributes to pass through the status we want to use. Otherwise the error servlet will set it to 500. So even though we were setting the error status, it was then getting overwritten.

We still do set the error status because in test setups forwarding doesn’t correctly work so it’s helpful to tell that the response is a 403.