We can't use the username on a user
At the moment when a user launches a LTI tool from a different placement a different subject (sub claim) is sent. This means we end up with multiple entries in the user repository for the same user. With the index on the username this caused problems.

While we could drop the user completely it's useful if we ever need to debug problems as it means we don't have to do the lookup from subject -> username.