Update the /api/run to return ContextJob
Having the ContextJob come back from the API means that we know the ID of the content job and so don't have to do any more calls and also we can just update the state in react with the new ContextJob.

It is also a common pattern to have the create method return the object normally worked with.