AB#57602 Workaround primary key conflicts
It appears we can get a job both fired and run at the same time. This results in the job progress object failing to save and so gets stuck. This doesn't fix the actual error but should mean the job does at least run. We still log the error, but when the job finishes it should correctly set the status to complete.