AB#61998 Save a timezone in the claims
When linking to a calendar we save the timezone of the user who is setting up the link as the timezone to perform all the imports with. This way we have a sensible timezone for all the imports that doesn't depend on the timezone of the user doing the import (which may well be different to that of the file being imported).

A better solution would have been to enforce the datetimes in the file being imported to include a timezone as then we wouldn't need this information. iCal is much better in this respect.

One further option would be to have the user setting up the URL to be able to specify the timezone.