AB#62758 Initial testing-library work
This tests the deep linking flow that is used when setting up an calendar import that can be used by users to import a calendar into their own personal calendar.

Gotchas so far, doesn't handle page navigation, this isn't much of a problem for most React stuff, but for the deep links we do actually need the navigation so we can't fully test it.

It also focuses on matching using the aria role which is good as it remains generic (should last through refactors), but means you need to understand them.

This also has coverage for UserCalendars.