AB#65008 Attempt to make tests more stable
We don't need the datasource defined in the application.properties any more as it restarts fine without it.

Quartz should be participating in transactions so that if we rollback something it rolls back the quartz changes.

We shouldn't need to mark everything as dirtying the context but I couldn't work out anything simpler in a reasonably amount of time.