AB#65081 Retry deployment tests
This is to improve the reliability of the deployment tests. We also add a message when there aren't any imports in the course and then update the deployment test so it works in any deployment of calendar import, rather than requiring specific data. This makes it easier to test locally.

I don't get false success messages with this, I still get test failures if I disable the calendar import backend and test it against my local deployment.

This also added a test of the new frontend message.