AB#65804 Add token refresh to accounts UI.
Because we weren't loading the sections when running in an account it didn't prompt the user to authorize the app. This meant that when they attempted to run the job it would fail because the proxy didn't have a token for them.

This extracts out the checking of the token into a new component and then wraps all the interface elements that need it.