AB#69228 Better documentation on deployment.
This refactors some of the configuration so that it's pulled from the LTI configuration instead of in the deployment. This makes it easier to automatically configure the tool as rather than editing more files it can be automatically substituted.

This also uses docker for the frontend so that it can all be deployed automatically.