AB#69228 Improve the deployment instructions.
I'd accidentally forgotten to include the application-docker.properties file which meant that it was using h2 instead of MySQL.

We also allow the issuer to be manually set as the hostname/port value isn't useful when running inside docker. The old way of using the hostname should still work fine.

This switches the instructions to use the newer lti-auto-configuration which prompts for configuration values.