AB#76646 Switch back to using a URL.
The main change is to switch back to using a URL as we need this for the subscription type imports which aren't stored locally. Technically it should be a URI, but the complexity of changing a field from url -> URL isn't worth it at the moment. I haven't checked but changing the parameters passed to the job would have meant that any existing subscriptions that were already setup would not have worked as the parameters needed for the job are set when the job is initially setup.

We split out the various ways of loading a file into multiple implementations of the DepositService. This allows a clearer way of handling the different ways of loading the files. There is then an implementation that asks each implementation in turn if they can handle a particular URI. Generally implementations do this by looking at the scheme(protocol).

The reason we should be talking about URIs and not URLs is that some of the values are not locators (we need more information to actually locate the file) and so it's better to think about them as URIs. An example is that the S3 bucket isn't in the URI as that means we don't need to change the URI from prod -> dev when we copy the DB.

One bug in the S3 implementation was that it used java.nio.Path to build keys for S3 buckets which works on Unix but on Windows I think this will break as the path separator is `\`.

The `?.toString` being appended to the expressions is to continue to support older Canvas version where numbers were sent across as integers in the claims and we had to have the parameter as an Object so that it could accept both Strings and Integers. A better way to handle this is to have a custom object for this data from the claims and parse the claims in one place.